Add dynamic routes for [group]
