+++ web/app/[locale]/(end-user)/[group]/_components/IndexTemplate.tsx
