+++ web/app/[locale]/(end-user)/[group]/channels/[id]/page.tsx
