+++ web/app/[locale]/(end-user)/[group]/charts/channels/_components/ChartTemplate.tsx
