+++ web/app/[locale]/(end-user)/[group]/charts/channels/page.tsx
@@ -1,21 +1,19 @@
 import { Metadata } from 'next'
 import { useTranslations } from 'next-intl'
 import { getTranslations, unstable_setRequestLocale } from 'next-intl/server'
-import { IndexTemplate } from 'app/[locale]/(end-user)/hololive/scheduled/_components/IndexTemplate'
+import { ChartTemplate } from 'app/[locale]/(end-user)/[group]/charts/channels/_components/ChartTemplate'
 import Page from 'components/Page'
 
 type Props = {
   params: { locale: string }
+  searchParams?: ConstructorParameters<typeof URLSearchParams>[0]
 }
 
 export async function generateMetadata({
   params: { locale }
 }: Props): Promise<Metadata> {
   const tg = await getTranslations({ locale, namespace: 'Global' })
-  const t = await getTranslations({
-    locale,
-    namespace: 'Page.hololive.scheduled.metadata'
-  })
+  const t = await getTranslations({ locale, namespace: 'Page.hololive.charts' })
 
   return {
     title: `${t('title')} | ${tg('title')}`,
@@ -23,7 +21,10 @@ export async function generateMetadata({
   }
 }
 
-export default function HololiveScheduledPage({ params: { locale } }: Props) {
+export default function HololiveChartsPage({
+  params: { locale },
+  searchParams
+}: Props) {
   // Enable static rendering
   unstable_setRequestLocale(locale)
 
@@ -33,10 +34,10 @@ export default function HololiveScheduledPage({ params: { locale } }: Props) {
     <Page
       breadcrumb={[
         { href: '/hololive', name: t('hololive') },
-        { href: '/hololive/scheduled', name: t('scheduled') }
+        { href: '/hololive/charts/channels', name: t('channels') }
       ]}
     >
-      <IndexTemplate />
+      <ChartTemplate />
     </Page>
   )
 }
