+++ web/app/[locale]/(end-user)/[group]/live/_components/IndexTemplate.tsx
