+++ web/app/[locale]/(end-user)/[group]/scheduled/_components/IndexTemplate.tsx
