+++ web/app/[locale]/(end-user)/[group]/scheduled/page.tsx
@@ -1,19 +1,21 @@
 import { Metadata } from 'next'
 import { useTranslations } from 'next-intl'
 import { getTranslations, unstable_setRequestLocale } from 'next-intl/server'
-import { ChartTemplate } from 'app/[locale]/(end-user)/hololive/charts/channels/_components/ChartTemplate'
+import { IndexTemplate } from 'app/[locale]/(end-user)/[group]/scheduled/_components/IndexTemplate'
 import Page from 'components/Page'
 
 type Props = {
   params: { locale: string }
-  searchParams?: ConstructorParameters<typeof URLSearchParams>[0]
 }
 
 export async function generateMetadata({
   params: { locale }
 }: Props): Promise<Metadata> {
   const tg = await getTranslations({ locale, namespace: 'Global' })
-  const t = await getTranslations({ locale, namespace: 'Page.hololive.charts' })
+  const t = await getTranslations({
+    locale,
+    namespace: 'Page.hololive.scheduled.metadata'
+  })
 
   return {
     title: `${t('title')} | ${tg('title')}`,
@@ -21,10 +23,7 @@ export async function generateMetadata({
   }
 }
 
-export default function HololiveChartsPage({
-  params: { locale },
-  searchParams
-}: Props) {
+export default function HololiveScheduledPage({ params: { locale } }: Props) {
   // Enable static rendering
   unstable_setRequestLocale(locale)
 
@@ -34,10 +33,10 @@ export default function HololiveChartsPage({
     <Page
       breadcrumb={[
         { href: '/hololive', name: t('hololive') },
-        { href: '/hololive/charts/channels', name: t('channels') }
+        { href: '/hololive/scheduled', name: t('scheduled') }
       ]}
     >
-      <ChartTemplate />
+      <IndexTemplate />
     </Page>
   )
 }
